added mac support
Mac support added (via using Wine). This differs from the Linux setup because Steam runs natively on Linux, but with proton configured to run the games.

On Mac, Wine runs Steam and then something like GPTK or CrossOver is used to play the game. (essentially the Mac setup is actually identical to the Windows setup)